move StylesComponents into styled.ts, mooved helper functions into new helper.ts module, and moved NewOpportunityDetailsFields into new module of same name
